LayoutTests/http/wpt/webauthn/public-key-credential-unfocused-document.https.html

<!DOCTYPE html>
<html>
<head>
    <meta charset="utf-8">
    <title>Web Authentication API: Tests that a frame that doesn't have the focus could not access the API.</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/common/utils.js"></script>
    <script src="/common/get-host-info.sub.js"></script>
    <script src="./resources/util.js"></script>
</head>
<body>
    <iframe src=""></iframe>
    <script>
        promise_test(t => {
            return withCrossOriginIframe("last-layer-frame.https.html?shouldFocus=false&exceptionMessage=The document is not focused.").then((message) => {
                assert_equals(message.data, "PASS.");
            });
        }, "Tests that a frame that doesn't have the focus could not access the API.");
    </script>
</body>
</html>
